+++ packages/spright-components/src/icons/breakpoint-execution-pointer/icon-data.ts
@@ -1,0 +1,11 @@
+export const iconData = `<?xml version="1.0" encoding="UTF-8"?>
+<svg id="Layer_1" data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16">
+  <defs>
+    <style>
+      .breakpoint-execution-pointer-0 {
+        fill: #161617;
+      }
+    </style>
+  </defs>
+  <polygon class="breakpoint-execution-pointer-0" points="7.97 4.37 7.29 5.49 9.8 7 2 7 2 9 9.8 9 7.29 10.51 7.97 11.63 14 8 7.97 4.37"/>
+</svg>`;