+++ packages/nimble-tokens/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-tokens
 
-<!-- This log was last generated on Mon, 04 May 2026 17:04:06 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 26 May 2026 18:42:52 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 8.18.1
+
+Tue, 26 May 2026 18:42:52 GMT
+
+### Patches
+
+- Update `nimble-icon-arrow-right-thin` ([ni/nimble@92b4050](https://github.com/ni/nimble/commit/92b405008ad93d76a4b80d8c3ce971f4f2cf6bc6))
 
 ## 8.18.0